Update matrix<bool> and mbool
